shortcut erase_marked_points when possible
